Revert and fix "Revert usage of sed back to rg"
This reverts commit 5e467cd045665a985dc612aa1ac48a7d51d8b73a.

Finally figured out what was going on: '\1' was being interpreted as the
literal character 1 in ASCII, which corresponds to a Ctrl-A, thus
explaining the '^A' I was seeing in the output (which I was
misinterpreting as being a literal character 10 (A is 10 in
hexadecimal), which is a newline and made me think the sed was doing
something funky).

Escaping the backslash on '\1' fixed the issue.